requirements.txt
faker==24.4.0
pandas==2.2.1
google-cloud-storage==2.16.0
google-cloud-bigquery==3.20.1
google-cloud-bigquery-storage==2.24.0
google-generativeai==0.5.4
streamlit==1.33.0
plotly==5.20.0
python-dotenv==1.0.1
pyarrow==15.0.2
db-dtypes==1.2.0
